refactor: remove unnecessary subpackages
