+++ pyquack/__init__.py
@@ -21,7 +21,7 @@ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 """
-from .response import QueryResult as QueryResult
+from .models import QueryResult as QueryResult
 from .sync import Client as Client
 
 try:
@@ -29,4 +29,4 @@ try:
 except ValueError:  # raised by _async when aiohttp is not present
     pass
 
-__version__ = "0.1.1"
+__version__ = "0.2.0"
