+++ pyquack/sync.py
@@ -23,11 +23,10 @@ SOFTWARE.
 """
 import requests
 
-from pyquack import QueryResult
-from pyquack.response import _parse_response
+from .models import _parse_response, QueryResult
 
 
-__version__ = "0.1.1"
+__version__ = "0.2.0"
 
 
 class Client:
